# open_api/openapi.yaml
openapi: 3.0.3
info:
  title: FormulOne
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: testing server
...
paths:
...
  /api/grandprix:
    get:
      tags:
        - grand_prix
      summary: Get all grand prix
      description: Returns all grand prix
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GrandPrixResponse'
        '500':
          description: Status internal server error

    post:
      tags:
        - grand_prix
      summary: Create grand prix
      description: ''
      requestBody:
        description: Create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrandPrix'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid input
        '401':
          description: Unauthorized user
        '403':
          description: Not enough rights
        '500':
          description: Status internal server error
  /api/grandprix/{id}:
    get:
      tags:
        - grand_prix
      summary: Get grand prix
      description: Return grand prix by id
      parameters:
        - name: id
          in: path
          description: ID of grand prix
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GrandPrixResponse'
        '400':
          description: Invalid input
        '404':
          description: Id doesnt exist
        '500':
          description: Status internal server error
    put:
      tags:
        - grand_prix
      summary: Update grand prix
      description: ''
      parameters:
        - name: id
          in: path
          description: ID of grand prix
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update grand prix
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrandPrix'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid input
        '401':
          description: Unauthorized user
        '403':
          description: Not enough rights
        '404':
          description: Id doesnt exist
        '500':
          description: Status internal server error
    delete:
      tags:
        - grand_prix
      summary: Delete grand prix
      description: ''
      parameters:
        - name: id
          in: path
          description: ID of grand prix
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
        '401':
          description: Unauthorized user
        '403':
          description: Not enough rights
        '404':
          description: Id doesnt exist
        '500':
          description: Status internal server error
  /api/grandprix/{id}/name:
    patch:
      tags:
        - grand_prix
      summary: Update grand prix name
      description: ''
      parameters:
        - name: id
          in: path
          description: ID of grand prix
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update grand prix name
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrandPrixPatch'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid input
        '401':
          description: Unauthorized user
        '403':
          description: Not enough rights
        '404':
          description: Id doesnt exist
        '500':
          description: Status internal server error
  /api/grandprix/{id}/race_results:
    get:
      tags:
        - grand_prix
      summary: Get race results of GP
      description: Return race results of GP by gp_id
      parameters:
        - name: id
          in: path
          description: ID of grand prix
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RaceResultView'
        '400':
          description: Invalid input
        '404':
          description: Id doesnt exist
        '500':
          description: Status internal server error
...
  /api/grandprix/{id}/qual_results:
    get:
      tags:
        - grand_prix
      summary: Get qual results of GP
      description: Return qual results of GP by gp_id
      parameters:
        - name: id
          in: path
          description: ID of grand prix
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QualResultView'
        '400':
          description: Invalid input
        '404':
          description: Id doesnt exist
        '500':
          description: Status internal server error
...
  /api/grandprix/{gp_id}/race_results/{id}:
    get:
      tags:
        - race_results
      summary: Get race result
      description: Return race result by id
      parameters:
        - name: gp_id
          in: path
          description: ID of gp
          required: true
          schema:
            type: integer
            format: int64
        - name: id
          in: path
          description: ID of race result
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RaceResultView'
        '400':
          description: Invalid input
        '404':
          description: Id doesnt exist
        '500':
          description: Status internal server error
    put:
      tags:
        - race_results
      summary: Update race result
      description: ''
      parameters:
        - name: gp_id
          in: path
          description: ID of gp
          required: true
          schema:
            type: integer
            format: int64
        - name: id
          in: path
          description: ID of race result
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update race result
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RaceResult'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid input
        '401':
          description: Unauthorized user
        '403':
          description: Not enough rights
        '404':
          description: Id doesnt exist
        '500':
          description: Status internal server error
    delete:
      tags:
        - race_results
      summary: Delete race result
      description: ''
      parameters:
        - name: gp_id
          in: path
          description: ID of gp
          required: true
          schema:
            type: integer
            format: int64
        - name: id
          in: path
          description: ID of race result
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
        '401':
          description: Unauthorized user
        '403':
          description: Not enough rights
        '404':
          description: Id doesnt exist
        '500':
          description: Status internal server error
...
components:
  schemas:
...
    GrandPrix:
      type: object
      properties:
        gp_id:
          type: integer
          format: int64
          example: 121
        gp_season:
          type: integer
          format: int64
          example: 2020
        gp_name:
          type: string
          example: Monaco GP
        gp_date_num:
          type: integer
          format: int64
          example: 23
        gp_month:
          type: string
          example: may
        gp_place:
          type: string
          example: Hungary
        gp_track_id:
          type: integer
          format: int64
          example: 2
    GrandPrixResponse:
      type: object
      properties:
        gp_id:
          type: integer
          format: int64
          example: 121
        gp_season:
          type: integer
          format: int64
          example: 2020
        gp_name:
          type: string
          example: Monaco GP
        gp_date_num:
          type: integer
          format: int64
          example: 23
        gp_month:
          type: string
          example: may
        gp_place:
          type: string
          example: Hungary
        gp_track_id:
          type: integer
          format: int64
          example: 2
        gp_featured:
          type: boolean
          example: true
    GrandPrixPatch:
      type: object
      properties:
        gp_name:
          type: string
          example: Monaco GP
    RaceResult:
      type: object
      properties:
        race_id:
          type: integer
          format: int64
          example: 121
        race_driver_place:
          type: integer
          format: int64
          example: 2
        driver_id:
          type: integer
          format: int64
          example: 25
        team_id:
          type: integer
          format: int64
          example: 4
        gp_id:
          type: integer
          format: int64
          example: 100
    RaceResultView:
      type: object
      properties:
        race_id:
          type: integer
          format: int64
          example: 121
        race_driver_place:
          type: integer
          format: int64
          example: 2
        driver_name:
          type: string
          example: Checo Perez
        team_name:
          type: string
          example: RedBull
        gp_name:
          type: string
          example: SPA
...
    QualResultView:
      type: object
      properties:
        qual_id:
          type: integer
          format: int64
          example: 121
        qual_driver_place:
          type: integer
          format: int64
          example: 2
        driver_name:
          type: string
          example: Checo Perez
        team_name:
          type: string
          example: RedBull
        q1_time:
          type: string
          example: '1:20.443'
        q2_time:
          type: string
          example: '1:20.125'
        q3_time:
          type: string
          example: '1:19.998'
        gp_name:
          type: string
          example: SPA